Refactor tests for databricks exceution engine
